+++ packages/api-shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @doccov/api-shared
 
+## 0.2.20
+
+### Patch Changes
+
+- @doccov/sdk@0.27.4
+
 ## 0.2.19
 
 ### Patch Changes